requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]==2.7.1
python-dotenv==1.0.1
google-auth==2.29.0
httpx==0.27.0
requests==2.31.0
groq==0.13.0